Rename files and test to be more consistent
