+++ meson.build
@@ -59,6 +59,7 @@ simde_neon_families = [
   'hadd',
   'hsub',
   'ld1',
+  'ld2',
   'ld3',
   'ld4',
   'max',
@@ -147,7 +148,6 @@ simde_neon_families = [
   'uzp1',
   'uzp2',
   'uzp',
-  'vld2',
   'xar',
   'zip1',
   'zip2',
