+++ simde/arm/neon.h
@@ -79,6 +79,7 @@
 #include "neon/hadd.h"
 #include "neon/hsub.h"
 #include "neon/ld1.h"
+#include "neon/ld2.h"
 #include "neon/ld3.h"
 #include "neon/ld4.h"
 #include "neon/max.h"
@@ -167,7 +168,6 @@
 #include "neon/uzp.h"
 #include "neon/uzp1.h"
 #include "neon/uzp2.h"
-#include "neon/vld2.h"
 #include "neon/xar.h"
 #include "neon/zip.h"
 #include "neon/zip1.h"
